django-rest/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
Django==4.1.5
djangorestframework==3.14.0
Faker==16.6.0
importlib-metadata==6.0.0
Markdown==3.4.1
pkg_resources==0.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
validate-docbr==1.10.0
zipp==3.11.0
